# backend/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "IT Help Desk Ticketing System API",
    "version": "1.0.0",
    "description": "Comprehensive API documentation for the IT Help Desk Ticketing Software Backend",
    "contact": {
      "name": "IT Helpdesk Team",
      "email": "support@ithelpdesk.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    },
    {
      "url": "http://localhost:5000",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT token in the format: Bearer <token>"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique user identifier",
            "example": "clp123abc456def789"
          },
          "username": {
            "type": "string",
            "description": "Username for login",
            "example": "john_doe"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "User email address",
            "example": "john.doe@company.com"
          },
          "role": {
            "type": "string",
            "enum": [
              "system_owner",
              "super_admin",
              "admin",
              "it_person",
              "user"
            ],
            "description": "User role in the system",
            "example": "user"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the user account is active",
            "example": true
          },
          "businessType": {
            "type": "string",
            "enum": ["small_business", "medium_business", "large_business"],
            "description": "Business type (for super_admin role)",
            "example": "medium_business"
          },
          "accountLimit": {
            "type": "integer",
            "description": "Account limit for super_admin",
            "example": 700
          },
          "expiryDate": {
            "type": "string",
            "format": "date-time",
            "description": "Account expiry date",
            "example": "2024-12-31T23:59:59Z"
          },
          "location": {
            "type": "string",
            "description": "User location",
            "example": "New York Office"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "User creation timestamp",
            "example": "2024-01-01T00:00:00Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "User last update timestamp",
            "example": "2024-01-01T00:00:00Z"
          }
        }
      },
...
      "ApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Indicates if the request was successful",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "Response message",
            "example": "Operation completed successfully"
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/users/create-super-admin": {
      "post": {
        "summary": "Create Super Admin account",
        "description": "Create a new Super Admin account with business type and account limits. Only System Owner can create Super Admins.",
        "tags": ["User Management"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "email",
                  "password",
                  "businessType",
                  "location"
                ],
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "Username for the Super Admin",
                    "example": "superadmin_company1"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Email address for the Super Admin",
                    "example": "superadmin@company.com"
                  },
                  "password": {
                    "type": "string",
                    "description": "Password for the Super Admin",
                    "example": "SecurePass123!"
                  },
                  "businessType": {
                    "type": "string",
                    "enum": [
                      "small_business",
                      "medium_business",
                      "large_business"
                    ],
                    "description": "Business type that determines account limits",
                    "example": "medium_business"
                  },
                  "expiryDate": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Account expiry date",
                    "example": "2025-12-31T23:59:59Z"
                  },
                  "location": {
                    "type": "string",
                    "description": "Super Admin location/office",
                    "example": "New York Headquarters"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Super Admin created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "user": {
                          "$ref": "#/components/schemas/User"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data or validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Only System Owner can create Super Admins",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict - Username or email already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/create-admin": {
      "post": {
        "summary": "Create Admin account",
        "description": "Create a new Admin account. Only Super Admins can create Admin accounts, subject to their account limits.",
        "tags": ["User Management"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "email", "password"],
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "Username for the Admin",
                    "example": "admin_dept1"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Email address for the Admin",
                    "example": "admin@company.com"
                  },
                  "password": {
                    "type": "string",
                    "description": "Password for the Admin",
                    "example": "SecurePass123!"
                  },
                  "ipNumber": {
                    "type": "string",
                    "description": "IP number for the Admin",
                    "example": "192.168.1.10"
                  },
                  "deviceName": {
                    "type": "string",
                    "description": "Device name",
                    "example": "ADMIN-LAPTOP-01"
                  },
                  "deviceIpAddress": {
                    "type": "string",
                    "description": "Device IP address",
                    "example": "192.168.1.10"
                  },
                  "location": {
                    "type": "string",
                    "description": "Admin location/department",
                    "example": "IT Department"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Admin created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "user": {
                          "$ref": "#/components/schemas/User"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data or validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Only Super Admins can create Admins or account limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict - Username or email already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/create-it-person": {
      "post": {
        "summary": "Create IT Person account",
        "description": "Create a new IT Person account. Admins and higher roles can create IT Person accounts.",
        "tags": ["User Management"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "email", "password"],
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "Username for the IT Person",
                    "example": "itperson_support1"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Email address for the IT Person",
                    "example": "itperson@company.com"
                  },
                  "password": {
                    "type": "string",
                    "description": "Password for the IT Person",
                    "example": "SecurePass123!"
                  },
                  "ipNumber": {
                    "type": "string",
                    "description": "IP number for the IT Person",
                    "example": "192.168.1.20"
                  },
                  "deviceName": {
                    "type": "string",
                    "description": "Device name",
                    "example": "IT-SUPPORT-01"
                  },
                  "deviceIpAddress": {
                    "type": "string",
                    "description": "Device IP address",
                    "example": "192.168.1.20"
                  },
                  "location": {
                    "type": "string",
                    "description": "IT Person location/department",
                    "example": "Help Desk - Floor 2"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "IT Person created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "user": {
                          "$ref": "#/components/schemas/User"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data or validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Only Admins and above can create IT Persons",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict - Username or email already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/users/{id}": {
      "get": {
        "summary": "Get user by ID",
        "description": "Retrieve a specific user by their ID. Users can only access their own profile or users they created.",
        "tags": ["User Management"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User ID",
            "example": "clp123abc456def789"
          }
        ],
        "responses": {
          "200": {
            "description": "User retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "user": {
                          "$ref": "#/components/schemas/User"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Cannot access this user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}